+++ src/features/board/ui/stackSearch/StackSearch.style.ts
@@ -3,4 +3,5 @@ import { styled } from 'styled-components';
 export const Container = styled.div`
   display: flex;
   gap: 2.4rem;
+  grid-column: 1/-1;
 `;
